92540 - ajuste end point
