+++ src/SME.SGP.Api/Controllers/ConselhoClasseController.cs
@@ -8,6 +8,7 @@ using SME.SGP.Dominio;
 using SME.SGP.Infra;
 using SME.SGP.Infra.Dtos;
 using SME.SGP.Infra.Dtos.ConselhoClasse;
+using SME.SGP.Infra.Dtos.Questionario;
 using SME.SGP.Infra.Dtos.Relatorios;
 using System.Collections.Generic;
 using System.Threading.Tasks;
@@ -282,6 +283,16 @@ namespace SME.SGP.Api.Controllers
         public async Task<IActionResult> AlterarParecerConclusivo(AlterarParecerConclusivoDto alterarParecerConclusivo, [FromServices] IAlterarParecerConclusivoUseCase useCase)
         {
             return Ok(await useCase.Executar(alterarParecerConclusivo));
+        }
+
+        [HttpGet("turmas/{codigoTurmaRegular}/alunos/{codigoAluno}/relatorios-pap")]
+        [ProducesResponseType(typeof(IEnumerable<SecaoQuestoesDTO>), 200)]
+        [ProducesResponseType(typeof(RetornoBaseDto), 500)]
+        [ProducesResponseType(typeof(RetornoBaseDto), 601)]
+        [Permissao(Permissao.RPAP_C, Permissao.CC_C, Policy = "Bearer")]
+        public async Task<IActionResult> ObterQuestionarioPAPConselhoClasse(string codigoTurmaRegular, string codigoAluno, int bimestre, [FromServices] IObterRelatorioPAPConselhoClasseUseCase useCase)
+        {
+            return Ok(await useCase.Executar(codigoTurmaRegular, codigoAluno, bimestre));
         }
     }
 }
